Installed, set up and modified code to use GSL numerical library.

+++ src/python/model_client/wrapper.py
@@ -11,7 +11,7 @@ model_core = ctypes.CDLL(dll_path)
 
 # Types for function signatures
 model_core.create_design_matrix.argtypes = [ctypes.c_int, ctypes.c_int]
-model_core.create_design_matrix.restype = ctypes.POINTER(ctypes.c_int)
+model_core.create_design_matrix.restype = ctypes.POINTER(ctypes.c_double)
 
 model_core.generate_synthetic_data.argtypes = [ctypes.c_double]
 model_core.generate_synthetic_data.restype = ctypes.c_double
